probadas distancia y maxQ, recompensa hace division entera

+++ src/AprendizajeQ.hpp
@@ -33,7 +33,7 @@ namespace AprendizajeQ {
 	
   void actualizarValor(int);
 	
-  int recompensa(int,int);
+  double recompensa(int,int);
   int distancia(int);
   int maxQ(int);
 }
